server/requirements.txt
google-genai
google-generativeai
flask
flask-cors
python-dotenv
wikipedia
requests
beautifulsoup4
gunicorn
pandas
numpy
packaging
pymongo
pyjwt
bcrypt
flask-limiter
PyPDF2
pdfplumber
